Changes from 0.8c: -Timer 2 is replaced with Timer 0 for step pulse duration (stepper.c) -The limit switch for the Z-axis is disabled -Entering the gcode command $L1 enables laser-mode and $L0 disables it -In laser mode, Z-axis commands are preserved at 0 to 255 values for setting a PWM pulse on Arduino Pin 11 (PORTB bit 3) to control laser intensity.  This PWM pulse can be fed directly into the modulation input of a laser driver like the FlexMod P3.  Then software like PicLaser Lite (picengrave.com) can be used to generate images with varying shades of gray/black.  A Z0 will have the laser at the lowest intensity set by the driver where a Z255 will set it to the maximum intensity/current the driver is set to allow.
Spindle Speed control commands were attempted first, but they are not
synchronized to axis movements.  By using the Z axis timing is improved
(though imperfect since PWM output control takes two or three more cycles
than is required to pulse stepper pins.)  Z axis usage is more compatible
with image engraving software.

A future addition may be another $L setting that is multiplier for Z value
so that 0.0 to 0.255 could be used with a multiplier of 1000, for better
compatibility with other software.